pyproject.toml
[tool.poetry]
name = "drf-twitter"
version = "0.1.0"
description = ""
authors = ["yu"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
flake8 = "^7.1.1"
Django = "4.2.9"
psycopg2 = "2.9.9"
psycopg2-binary = "2.9.9"
dj-database-url = "0.5.0"
djangorestframework = "3.14.0"
pytest = "8.3.3"
black = "24.10.0"
djangorestframework-simplejwt = "5.3.1"
drf-spectacular = "0.27.2"
django-filter = "24.3"
pytest-cov = "5.0.0"
django-environ = "^0.11.2"
dj-rest-auth = "5.0.1"
pytest-srcpaths = "1.2.1"
pytest-django = "4.9.0"
poetry = "1.8.0"
google-auth = "2.36.0"
google-cloud-secret-manager = "2.21.0"
gunicorn = "23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mysite.settings.local"
pythonpath = ["src/mysite"]